# src/backend/product/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/product": {
            "post": {
                "description": "Create product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create product",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_models_products.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_models_products.CreateProductResponse"
                        }
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "description": "Get product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/product-service_internal_application_queries_products.GetProductQueryByIdResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "product-service_internal_application_queries_products.GetProductQueryByIdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "product-service_internal_models_products.CreateProductRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string",
                    "example": "phones"
                },
                "name": {
                    "type": "string",
                    "example": "iPhone 13"
                },
                "price": {
                    "type": "number",
                    "example": 999.99
                },
                "stock": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "product-service_internal_models_products.CreateProductResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        }
    }
}